requirements.txt
torch
torchvision
pillow
tensorboard